Jdk15 (#7)

+++ doctor-core/src/main/java/vest/doctor/scheduled/Scheduled.java
@@ -13,7 +13,7 @@ import java.util.concurrent.TimeUnit;
  * Marks a method to be automatically scheduled with the default {@link java.util.concurrent.ScheduledExecutorService}
  * provided by the {@link ProviderRegistry}.
  * <p>
- * One and only one of {@link #interval()} or {@link #cron()} must be set.
+ * Only one of {@link #interval()} or {@link #cron()} may be set.
  */
 @Documented
 @Retention(RetentionPolicy.RUNTIME)
@@ -50,5 +50,5 @@ public @interface Scheduled {
      * Limit the number of times the scheduled method will execute per-instance of the containing object.
      * When negative, indicates that the method will be scheduled indefinitely.
      */
-    int executionLimit() default -1;
+    long executionLimit() default -1;
 }
